Extract message backend classes into separate files
